// apps/web/src/grpc/generated/pdf.ts
export interface ProtoGrpcType {
  Content: MessageTypeDefinition;
  SectionContent: MessageTypeDefinition;
  TextSource: EnumTypeDefinition;
  google: {
    protobuf: {
      Any: MessageTypeDefinition;
      Empty: MessageTypeDefinition;
      Timestamp: MessageTypeDefinition;
    };
  };
  learningmotors: {
    pb: {
// ...
      pdf: {
// ...
        PDF: MessageTypeDefinition;
        PDFService: SubtypeConstructor<typeof grpc.Client, _learningmotors_pb_pdf_PDFServiceClient> & {
          service: _learningmotors_pb_pdf_PDFServiceDefinition;
        };
        ZIP: MessageTypeDefinition;
      };
// ...
    };
  };
// ...
}