// apps/web/src/grpc/generated/sms.ts
type SubtypeConstructor<Constructor extends new (...args: any) => any, Subtype> = {
  new (...args: ConstructorParameters<Constructor>): Subtype;
};

export interface ProtoGrpcType {
// ...
  suki: {
    pb: {
// ...
      sms: {
// ...
        SMS: SubtypeConstructor<typeof grpc.Client, _suki_pb_sms_SMSClient> & { service: _suki_pb_sms_SMSDefinition };
// ...
      };
    };
  };
}